requirements.txt
asgiref==3.7.2
Django==4.2.2
django-appconf==1.0.5
django-compressor==4.3.1
django-sass-processor==1.2.2
libsass==0.22.0
python-dateutil==2.8.2
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
